src/layout/: init, moved Appbar and Navbar
